requirements.txt
flask
google-cloud-storage
google-cloud-bigquery
temp
gunicorn
google-auth
authlib